requirements.txt
yfinance
pandas
numpy
scikit-learn
streamlit
matplotlib
tensorflow-macos
